Modify DefaultTokenPainter.paintTabLines() and TokenImpl.getWidthUpTo() to use FPAPI
Refactor: move code from ConfigurableCaret.paint() to SwingUtils.drawLine()

+++ RSyntaxTextArea/src/main/java/org/fife/ui/rsyntaxtextarea/VisibleWhitespaceTokenPainter.java
@@ -164,7 +164,7 @@ public class VisibleWhitespaceTokenPainter extends DefaultTokenPainter {
 		// other than Token.WHITESPACE for spaces (such as Token.IDENTIFIER).
 		// This also allows us to paint tab lines for MLC's.
 		if (host.getPaintTabLines() && origX==host.getMargin().left) {// && isWhitespace()) {
-			paintTabLines(token, origX, (int)y, (int)nextX, g, e, host);
+			paintTabLines(token, origX, y, nextX, g, e, host);
 		}
 
 		return nextX;
